+++ utils.h
@@ -1,11 +1,16 @@
 #ifndef UTILS_H
 #define UTILS_H
-char* utils_get_login(void);
-char* utils_get_user_home(void);
-char* utils_get_local_apps_dir(void);
-char* utils_get_local_icons_dir(void);
-char* utils_get_cache_path(void);
-char* utils_get_logo_path(void);
-char* utils_get_wallpaper_path(void);
+
+typedef enum utils_t {
+    UTILS_LOGIN,
+    UTILS_USER_HOME,
+    UTILS_LOCAL_APPS,
+    UTILS_LOCAL_ICONS,
+    UTILS_CACHE,
+    UTILS_LOGO,
+    UTILS_WALLPAPER
+} utils_t;
+
+char* utils_get(utils_t util);
 int utils_is_root(void);
 #endif